FE: Add Loading Table
