+++ frontend/app/questions/loading.tsx
@@ -1,0 +1,7 @@
+import LoadingTable from "./Loading/LoadingTable";
+
+export default function Loading() {
+    return (
+        <LoadingTable />
+    )
+}
